cofirm minor service changes

+++ src/main/java/ru/noleg/crmmm/repository/GroupRepository.java
@@ -3,5 +3,5 @@ package ru.noleg.crmmm.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.noleg.crmmm.repository.model.Group;
 
-public interface GroupRepository extends CrudRepository<Group, Integer> {
+public interface GroupRepository extends CrudRepository<Group, Long> {
 }
